+++ tpcwithdnn/symmetry_padding_3d.py
